requirements.txt
streamlit
faiss-cpu
transformers
sentence-transformers
PyMuPDF
torch
google.generativeai
numpy
tqdm
kaggle